--- app.rb
@@ -1,3 +1,0 @@
-def ounces_to_grams(ounces)
-  ounces.to_f * 28.3495
-end
