--- app_test.rb
@@ -1,9 +1,0 @@
-require_relative './app.rb'
-
-puts "Given 1, it should return 28.3495"
-answer = ounces_to_grams(1)
-if answer == 28.3495
-  puts "Passed!"
-else
-  puts "Test failed: got #{answer} instead"
-end
