+++ conversions.rb
@@ -1,0 +1,4 @@
+GRAMS_PER_OUNCE = 28.3495
+def ounces_to_grams(ounces)
+  ounces.to_f * GRAMS_PER_OUNCE
+end
